feat: pull out QuickSelect and add organism and tissue select
